requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0
python-frontmatter>=1.1.0
